requirements.txt
flask==3.0.3
Flask-Migrate==4.0.7
flask-sqlalchemy==3.1.1
